Separate all unit tests into 3 files
